+++ src/radarChart.ts
@@ -60,6 +60,7 @@ import VisualUpdateOptions = powerbi.extensibility.visual.VisualUpdateOptions;
 import IVisualHost = powerbi.extensibility.visual.IVisualHost;
 import VisualTooltipDataItem = powerbi.extensibility.VisualTooltipDataItem;
 import ILocalizationManager = powerbi.extensibility.ILocalizationManager;
+import IVisualEventService = powerbi.extensibility.IVisualEventService;
 
 // Svg utils
 import * as SvgUtils from "powerbi-visuals-utils-svgutils";
@@ -212,6 +213,7 @@ export class RadarChart implements IVisual {
     private behavior: IInteractiveBehavior;
     private visualHost: IVisualHost;
     private localizationManager: ILocalizationManager;
+    private events: IVisualEventService;
 
     private tooltipServiceWrapper: ITooltipServiceWrapper;
 
@@ -451,6 +453,7 @@ export class RadarChart implements IVisual {
         this.formattingSettingsService = new FormattingSettingsService(this.localizationManager);
         this.interactivityService = createInteractivityService(this.visualHost);
         this.behavior = new RadarChartWebBehavior();
+        this.events = options.host.eventService;
 
         this.tooltipServiceWrapper = createTooltipServiceWrapper(
             options.host.tooltipService,
@@ -489,6 +492,7 @@ export class RadarChart implements IVisual {
             this.clear();
             return;
         }
+        this.events.renderingStarted(options);
         const dataView: DataView = options.dataViews[0];
 
         this.formattingSettings = RadarChart.parseSettings(dataView, this.colorHelper, this.formattingSettingsService);
@@ -516,6 +520,7 @@ export class RadarChart implements IVisual {
             categories = dataView.categorical.categories[0].values;
         } else {
             this.clear();
+            this.events.renderingFinished(options);
             return;
         }
         this.viewport = {
@@ -552,6 +557,7 @@ export class RadarChart implements IVisual {
 
         if ((width < RadarChart.MinViewportToRender.width) || (height < RadarChart.MinViewportToRender.height)) {
             this.clear();
+            this.events.renderingFinished(options);
             return;
         }
 
@@ -568,6 +574,7 @@ export class RadarChart implements IVisual {
 
         this.createAxesLabels();
         this.drawChart(series, RadarChart.AnimationDuration);
+        this.events.renderingFinished(options);
     }
 
     public getMinValue(dataView: DataView) : number {